model/main/requirements.txt
pandas
numpy
tensorflow
pillow
imageio
keras